requirements.txt
PyMuPDF==1.23.7
ollama
ruff==0.3.4 # Linter/Formatter opcional pero recomendado
PyPDF2
crewai>=0.1.0
requests
datadog
elasticsearch
python-dotenv
pytest

